[JSC] BBQ materializePointer should take uint64_t offset https://bugs.webkit.org/show_bug.cgi?id=321043 rdar://184072653
Reviewed by Keith Miller.

For 64bit address, we first do bound-check and after that, we
materialize a pointer. Currently since we are not supporting memory size
more than 4GB, this truncation does not matter, but we should have
uint64_t as an offset when we really want to have 64bit pointer. This
patch fixes it. Note that we never have overflow by adding 64bit offset
here since we already did bound-check and that already guaranteed this
pointer's validity. We also fixes recordJumpToThrowException cases to
correctly drain a pointer and returning garbage register even though we
are throwing an error.

* Source/JavaScriptCore/wasm/WasmBBQJIT.cpp:
(JSC::Wasm::BBQJITImpl::BBQJIT::materializePointer):
* Source/JavaScriptCore/wasm/WasmBBQJIT.h:
(JSC::Wasm::BBQJITImpl::BBQJIT::emitCheckAndPreparePointer):
* Source/JavaScriptCore/wasm/WasmBBQJIT64.h:
(JSC::Wasm::BBQJITImpl::BBQJIT::emitCheckAndPrepareAndMaterializePointerApply):

Canonical link: https://commits.webkit.org/318606@main

+++ Source/JavaScriptCore/wasm/WasmBBQJIT.cpp
@@ -1122,10 +1122,10 @@ void BBQJIT::emitMutatorFence()
 
 // Memory
 
-Address BBQJIT::materializePointer(Location pointerLocation, uint32_t uoffset)
+Address BBQJIT::materializePointer(Location pointerLocation, uint64_t uoffset)
 {
-    if (static_cast<uint64_t>(uoffset) > static_cast<uint64_t>(std::numeric_limits<int32_t>::max()) || !B3::Air::Arg::isValidAddrForm(B3::Air::Move, static_cast<int32_t>(uoffset), Width::Width128)) {
-        m_jit.addPtr(TrustedImmPtr(static_cast<int64_t>(uoffset)), pointerLocation.asGPR());
+    if (uoffset > static_cast<uint64_t>(std::numeric_limits<int32_t>::max()) || !B3::Air::Arg::isValidAddrForm(B3::Air::Move, static_cast<int32_t>(uoffset), Width::Width128)) {
+        m_jit.addPtr(TrustedImmPtr(uoffset), pointerLocation.asGPR());
         return Address(pointerLocation.asGPR());
     }
     return Address(pointerLocation.asGPR(), static_cast<int32_t>(uoffset));
